# api/openapi.yaml
openapi: 3.1.0
info:
  title: NimbusDB API
  version: 1.0.0-phase1
  description: |
    Control-plane REST API for the NimbusDB serverless PostgreSQL platform.
    Spec-first contract (ADR-012): handlers and clients are generated/validated from this file.
    Phase 1 surface: bootstrap, orgs, members, API keys, project records, audit log.
    Data-plane resources (branches, endpoints, backups, imports) arrive in Phase 2+ per
    docs/architecture/API_SPECIFICATION.md.
servers:
  - url: https://api.db.nimbus.app/v1
  - url: http://localhost:8080/v1
    description: Local development
security:
  - apiKey: []
...
paths:
...
  /bootstrap:
    post:
      tags: [system]
      operationId: bootstrap
      security: []
      summary: One-time platform bootstrap (ADR-013)
      description: |
        Creates the initial user, organization, and owner API key. Requires the
        `NDB_BOOTSTRAP_TOKEN` configured on the server and succeeds only while no
        organization exists. The returned API key is shown exactly once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bootstrap_token: { type: string }
                email: { type: string, format: email }
                name: { type: string }
                org_name: { type: string }
              required: [bootstrap_token, email, org_name]
      responses:
        "201":
          description: Platform bootstrapped
          content:
            application/json:
              schema:
                type: object
                properties:
                  org: { $ref: "#/components/schemas/Org" }
                  user: { $ref: "#/components/schemas/User" }
                  api_key: { $ref: "#/components/schemas/ApiKeyWithSecret" }
                required: [org, user, api_key]
        "403": { $ref: "#/components/responses/Problem" }
        "409":
          description: Platform already bootstrapped
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
  /orgs/{org}/members:
    parameters:
      - $ref: "#/components/parameters/org"
    get:
      tags: [orgs]
      operationId: listMembers
      summary: List organization members
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Member" }
                required: [data]
    post:
      tags: [orgs]
      operationId: addMember
      summary: Add an organization member
      description: Adds (or invites, once console auth exists in Phase 3) a member by email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                name: { type: string, description: Display name used if the user record is created by this call }
                role: { $ref: "#/components/schemas/OrgRole" }
              required: [email, role]
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "409": { $ref: "#/components/responses/Problem" }
  /orgs/{org}/members/{user}:
    parameters:
      - $ref: "#/components/parameters/org"
      - name: user
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [orgs]
      operationId: updateMemberRole
      summary: Change a member's role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: "#/components/schemas/OrgRole" }
              required: [role]
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "409":
          description: Would remove the last owner
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
    delete:
      tags: [orgs]
      operationId: removeMember
      summary: Remove an organization member
      responses:
        "204": { description: Removed }
        "409":
          description: Would remove the last owner
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /orgs/{org}/api-keys:
    parameters:
      - $ref: "#/components/parameters/org"
    get:
      tags: [api-keys]
      operationId: listApiKeys
      summary: List API keys
      responses:
        "200":
          description: API keys (metadata only; secrets are never re-shown)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ApiKey" }
                required: [data]
    post:
      tags: [api-keys]
      operationId: createApiKey
      summary: Create an API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                scopes:
                  type: array
                  items: { $ref: "#/components/schemas/Scope" }
                  minItems: 1
                expires_at: { type: [string, "null"], format: date-time }
              required: [name, scopes]
      responses:
        "201":
          description: Created key — secret shown exactly once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyWithSecret" }
  /orgs/{org}/api-keys/{key}:
    parameters:
      - $ref: "#/components/parameters/org"
      - name: key
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [api-keys]
      operationId: revokeApiKey
      summary: Revoke an API key
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/Problem" }
...
  /projects:
    get:
      tags: [projects]
      operationId: listProjects
      summary: List projects
      parameters:
        - name: org
          in: query
          required: false
          schema: { type: string }
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
                  next_cursor: { type: [string, "null"] }
                required: [data]
    post:
      tags: [projects]
      operationId: createProject
      summary: Create a project
      description: |
        Phase 1: creates the project *record* (state `pending`); data-plane provisioning
        attaches in Phase 2 via the reconciler.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                org_id: { type: string }
                name: { type: string, minLength: 1, maxLength: 120 }
                region: { type: string, default: syd1 }
                pg_version: { type: integer, enum: [16, 17], default: 17 }
              required: [org_id, name]
      responses:
        "201":
          description: Created project record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "409": { $ref: "#/components/responses/Problem" }
  /projects/{prj}:
    parameters:
      - name: prj
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [projects]
      operationId: getProject
      summary: Get a project
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [projects]
      operationId: updateProject
      summary: Update a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      tags: [projects]
      operationId: deleteProject
      summary: Delete a project
      responses:
        "204": { description: Deleted (soft; lifecycle per MULTI_TENANCY §5) }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: "ndb_<64 hex chars>"
  parameters:
    org:
      name: org
      in: path
      required: true
      schema: { type: string }
      description: Organization ID (`org_…`)
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        request_id: { type: string }
      required: [title, status]
    OrgRole:
      type: string
      enum: [owner, admin, member, viewer]
    Scope:
      type: string
      enum:
        - orgs:read
        - projects:read
        - projects:write
        - projects:provision
        - branches:read
        - branches:write
        - endpoints:read
        - roles:read
        - roles:write
        - backups:read
        - backups:write
        - restores:write
        - imports:read
        - imports:write
        - metrics:read
        - audit:read
        - keys:manage
        - webhooks:manage
        - usage:read
    Org:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [free, pro] }
        created_at: { type: string, format: date-time }
      required: [id, name, slug, plan, created_at]
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
      required: [id, email, created_at]
    Member:
      type: object
      properties:
        org_id: { type: string }
        user: { $ref: "#/components/schemas/User" }
        role: { $ref: "#/components/schemas/OrgRole" }
        added_at: { type: string, format: date-time }
      required: [org_id, user, role, added_at]
    ApiKey:
      type: object
      properties:
        id: { type: string }
        org_id: { type: string }
        name: { type: string }
        prefix:
          type: string
          description: First 12 characters of the token, for identification.
        scopes:
          type: array
          items: { $ref: "#/components/schemas/Scope" }
        last_used_at: { type: [string, "null"], format: date-time }
        expires_at: { type: [string, "null"], format: date-time }
        revoked_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
      required: [id, org_id, name, prefix, scopes, created_at]
    ApiKeyWithSecret:
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          properties:
            token:
              type: string
              description: Full secret token. Returned exactly once; store it now.
          required: [token]
    Project:
      type: object
      properties:
        id: { type: string }
        org_id: { type: string }
        name: { type: string }
        slug: { type: string }
        region: { type: string }
        pg_version: { type: integer }
        state:
          type: string
          enum: [pending, provisioning, ready, error, deleting]
          description: Phase 1 records stay `pending`; the Phase 2 reconciler drives the rest.
        default_branch_id: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
      required: [id, org_id, name, slug, region, pg_version, state, default_branch_id, created_at]
...